+++ docs/tr1/CHANGELOG.md
@@ -10,7 +10,7 @@
 - added ingame help for all settings
 - added support for object name aliases; added aliases for dev commands
 - added an optional breeze effect for Lara's braid in appropriate outside rooms (#3090)
-- added key icons to the controls settings dialog
+- added keyboard and controller input icons to the controls settings dialog
 - added an option to adjust music and ambient volume while in inventory
 - added a `/secret` console command for easier debugging of secrets
 - added `enable_debug_pos` setting that shows Lara's position in realtime (reachable via `/debug`, fine-tuned `/set`)